+++ R/sanitize_hypothesis.R
@@ -1,5 +1,9 @@
 sanitize_hypothesis <- function(hypothesis, ...) {
 
+    if (isTRUE(checkmate::check_formula(hypothesis))) {
+        hypothesis <- sanitize_hypothesis_formula(hypothesis)
+    }
+
     checkmate::assert(
         checkmate::check_character(hypothesis, pattern = "="),
         checkmate::check_choice(hypothesis, choices = c("pairwise", "reference", "sequential", "meandev", "meanotherdev", "revpairwise", "revreference", "revsequential")),